@@ -5,32 +5,24 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=7.1.3",
+        "php": ">=8.0.2",
         "ext-curl": "*",
         "ext-json": "*",
         "doctrine/dbal": "^2.9",
-        "fideloper/proxy": "^4.0",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^9.0",
         "laravel/helpers": "^1.4",
         "laravel/tinker": "^2.0",
         "laravelcollective/html": "^6.2",
-        "markbaker/complex": "1.5.0",
         "mk-j/php_xlsxwriter": "^0.38.0",
-        "phpoffice/phpspreadsheet": "1.17.1",
         "phpxmlrpc/phpxmlrpc": "4.*",
         "promodo/laravel-azure-auth": "^1.0",
-        "rap2hpoutre/fast-excel": "^1.2",
-        "rootinc/laravel-azure-middleware": "^0.5.1",
         "sentry/sentry-laravel": "^2.10",
         "symfony/dom-crawler": "^5.2"
     },
     "require-dev": {
         "filp/whoops": "~2.0",
-        "fzaninotto/faker": "~1.4",
-        "laravel/dusk": "^5.0",
         "mockery/mockery": "~1.0",
-        "mpociot/laravel-apidoc-generator": "^4.8",
-        "phpunit/phpunit": "^8.5",
+        "phpunit/phpunit": "^9.0",
         "symfony/thanks": "^1.0"
     },
     "autoload": {
